% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_anchors.R
\name{build_anchors}
\alias{build_anchors}
\title{build HTML anchors for pubmed ids}
\usage{
build_anchors(pmids)
}
\arguments{
\item{pmids}{character() vector (maybe numeric) of pubmed ids}
}
\value{
anchors in a character vector
}
\description{
build HTML anchors for pubmed ids
}
\examples{
build_anchors(c("12345", "23456"))
}
