% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curated_scrnaseq_table}
\alias{curated_scrnaseq_table}
\title{scRNAseq datasets in an annotated table}
\format{
data.frame
}
\usage{
curated_scrnaseq_table
}
\description{
scRNAseq datasets in an annotated table
}
\keyword{datasets}
